feedback part 1

+++ src/main/java/de/tum/in/www1/artemis/service/FileUploadAssessmentService.java
@@ -75,7 +75,9 @@ public class FileUploadAssessmentService extends AssessmentService {
         }
 
         // remove result from TutorScore because the assessor changes here
-        tutorScoreService.removeResult(result);
+        if (result.getScore() != null) {
+            tutorScoreService.removeResult(result);
+        }
 
         result.setHasComplaint(false);
         result.setExampleResult(fileUploadSubmission.isExampleSubmission());
