+++ src/main/java/de/tum/in/www1/artemis/service/TextAssessmentService.java
@@ -85,7 +85,9 @@ public class TextAssessmentService extends AssessmentService {
         Result result = desiredResult.orElseGet(Result::new);
 
         // remove result from TutorScore because the assessor changes here
-        tutorScoreService.removeResult(result);
+        if (result.getScore() != null) {
+            tutorScoreService.removeResult(result);
+        }
 
         User user = userService.getUser();
         result.setAssessor(user);
